requirements.txt
aiohttp==3.13.3
asyncudp==0.11.0
blinker==1.9.0
ezdxf==1.3.5
GitPython==3.1.44
numpy==2.3.4
opencv_python
platformdirs==4.3.6
pluggy==1.6.0
pycairo==1.28.0
pyclipper==1.3.0.post6
PyGObject==3.50.0
PyOpenGL==3.1.9
PyOpenGL_accelerate==3.1.9
pypdf==6.6.0
pyserial_asyncio==0.6
pyvips==3.0.0
PyYAML==6.0.2
scipy==1.16.3
semver==3.0.2
svgelements==1.9.6
vtracer==0.6.11
websockets==14.2
